PKI code modularized
